Move functions to schema_utils to break circular dependency
